tpvCommerce/apps_tpvCommerce/user/api/urls.py
""" Registro urls """

# Django REST Framework
from rest_framework import routers

# Django
from django.conf.urls import url

# Views
from apps_tpvCommerce.user.api.views import UserViewSet

router = routers.DefaultRouter()
router.register(r'^api/v1/user', UserViewSet)

urlpatterns = router.urls
