tpvCommerce/requirements.txt
Django
django-filter
psycopg2
djangorestframework
#psycopg2-binary